+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - basic autodownloading of arXiv PDF files (#25,!35):
     - the default download location can be configured via `config.utils.file_downloader.default_location`
     - on a per addition basis, this default can be overwritten via the `--path` keyword option of the `AddCommand`
+    - if a file already exists in this location of the file system, the download will be skipped
 - automatic journal abbreviations (#62,!36):
     - users can configure a list of journal abbreviations via `config.utils.journal_abbreviations`
     - if present, coBib will store the journal in its elongated form